Introduce fully functional 'truncate' system call
Previously, the underlying file system ext2 only supported truncate up
to the 0 size. It means that we could only wipe the file's content out.
But we want to fully support the 'truncate' system call so the
implementation has been developed.
Note: this version only supports the truncate to the less value than the file
size itself;