Fix the bug in `ext2_readdir` which was caused by leaking the buffer cache entry
Viewing `lost+found` directory causes the kernel to panic.
The kernel panicked with `[bcache]: no buffers`. This signified that
somewhere we leaked the buffer cache entries. In the case of repetitive
iterating over ext2 directory entries which are empty (which is the case
in `lost+found`) we forget to the release the buffer cache entry after
examining its contents.
The solution to this is pretty simple: release buffer cache entry after
examining with it.