signals: handle all pending signals when returning to the user space
Previously the kernel handled only 1 signal in check_signals() function.
Now we change the behaviour to handle all pending signals.
This is necessary because the signal handling may block the process and
returning from the function handle_signal() we can have a pending signal
(because we were awoken up due to another signal e.g. SIGCONT).